# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/rooms/create": {
            "post": {
                "description": "Добавляет новую комнату",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rooms"
                ],
                "summary": "Создать комнату",
                "parameters": [
                    {
                        "description": "Данные комнаты",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internship_internal_models_dto_handler.CreateRoomRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internship_internal_models_dto_handler.CreateRoomResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_http-server_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_http-server_handler.InternalErrorResponse"
                        }
                    }
                }
            }
        },
        "/rooms/list": {
            "get": {
                "description": "Возвращает список всех доступных комнат",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rooms"
                ],
                "summary": "Список комнат",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internship_internal_models_dto_handler.ListRoomsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_http-server_handler.InternalErrorResponse"
                        }
                    }
                }
            }
        },
        "/rooms/{roomID}/schedule/create": {
            "post": {
                "description": "Задает график работы для комнаты",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rooms"
                ],
                "summary": "Создать расписание",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID комнаты",
                        "name": "roomID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные расписания",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internship_internal_models_dto_handler.CreateScheduleRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internship_internal_models_dto_handler.CreateScheduleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_http-server_handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_http-server_handler.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/internal_http-server_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_http-server_handler.InternalErrorResponse"
                        }
                    }
                }
            }
        },
        "/rooms/{roomID}/slots/list": {
            "get": {
                "description": "Возвращает свободные слоты для комнаты на конкретную дату",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rooms"
                ],
                "summary": "Доступные слоты",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID комнаты",
                        "name": "roomID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Дата в формате YYYY-MM-DD",
                        "name": "date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internship_internal_models_dto_handler.ListSlotsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_http-server_handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_http-server_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_http-server_handler.InternalErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "internal_http-server_handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "internal_http-server_handler.InternalErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "internship_internal_models_dto_handler.CreateRoomRequest": {
            "type": "object",
            "properties": {
                "capacity": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "internship_internal_models_dto_handler.CreateRoomResponse": {
            "type": "object",
            "properties": {
                "room": {
                    "$ref": "#/definitions/internship_internal_models_dto_handler.RoomData"
                }
            }
        },
        "internship_internal_models_dto_handler.CreateScheduleRequest": {
            "type": "object",
            "properties": {
                "daysOfWeek": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "endTime": {
                    "type": "string"
                },
                "startTime": {
                    "type": "string"
                }
            }
        },
        "internship_internal_models_dto_handler.CreateScheduleResponse": {
            "type": "object",
            "properties": {
                "schedule": {
                    "$ref": "#/definitions/internship_internal_models_dto_handler.ScheduleData"
                }
            }
        },
...
        "internship_internal_models_dto_handler.ListRoomsResponse": {
            "type": "object",
            "properties": {
                "rooms": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internship_internal_models_dto_handler.RoomData"
                    }
                }
            }
        },
        "internship_internal_models_dto_handler.ListSlotsResponse": {
            "type": "object",
            "properties": {
                "slots": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internship_internal_models_dto_handler.SlotData"
                    }
                }
            }
        },
...
        "internship_internal_models_dto_handler.RoomData": {
            "type": "object",
            "properties": {
                "capacity": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "internship_internal_models_dto_handler.ScheduleData": {
            "type": "object",
            "properties": {
                "daysOfWeek": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "endTime": {
                    "type": "string"
                },
                "roomId": {
                    "type": "string"
                },
                "startTime": {
                    "type": "string"
                }
            }
        },
        "internship_internal_models_dto_handler.SlotData": {
            "type": "object",
            "properties": {
                "end": {
                    "type": "string"
                },
                "roomId": {
                    "type": "string"
                },
                "start": {
                    "type": "string"
                }
            }
        },
...
    }
}